fix(#14): address PR #69 review comments
- Remove ParsedTicket alias, use IssueTemplate directly
- Extract duplicated readEnv() to shared src/utils/env.ts
- Move CONTRIBUTING.md from src/ to designs/
- Move test file to test/layers/code-gen/ matching src/ convention
- Merge duplicate imports in test file
- Add missing test cases (env var fallback, alias resolution, empty inputs)
- Revert unrelated proxyquire change in context-loader.spec.ts
- Regenerate clean package-lock.json removing spurious peer additions
