+++ src/utils/env.ts
@@ -1,0 +1,3 @@
+export function readEnv(name: string): string | undefined {
+  return process.env[name];
+}
