+++ test/utils/context-loader.spec.ts
@@ -1,5 +1,4 @@
 import { expect } from 'chai';
-import proxyquireImport from 'proxyquire';
 import {
   parseFrontmatter,
   loadDomainOverview,
@@ -9,7 +8,8 @@ import {
   ensureAgentMemoryExists,
 } from '../../src/utils/context-loader';
 
-const proxyquire = proxyquireImport.noCallThru();
+// Use require for proxyquire to avoid ESM conflicts
+const proxyquire = require('proxyquire').noCallThru();
 
 describe('context-loader', () => {
   describe('parseFrontmatter', () => {
